[FIX] Fixing bad/buggy code, removing prints.

+++ skwdro/linear_models/_logistic_regression.py
@@ -148,8 +148,8 @@ class LogisticRegression(BaseEstimator, ClassifierMixin):
         self.le_ = LabelEncoder()
         y = self.le_.fit_transform(y)
         if y is None: raise ValueError("Problem with labels, none out of label encoder")
-        else: y = np.array(y)
-        y[y==0] = -1
+        else: y = np.array(y, dtype=X.dtype)
+        y[y==0.] = -1.
 
         if len(self.classes_)>2:
             raise NotImplementedError(f"Multiclass classificaion is not implemented. ({len(self.classes_)} classes were found : {self.classes_})")
@@ -245,6 +245,10 @@ class LogisticRegression(BaseEstimator, ClassifierMixin):
             
             # # TODO: deprecate ?
             # # Stock the robust loss result
+            # Problems w/ dtypes (f32->f64 for some reason)
+            # To be fixed later
+            # =============================================
+            # Stock the robust loss result
             # if self.solver == "entropic_torch_pre":
             #     #self.result_ = self.problem_.loss.forward(xi=self.X_, xi_labels=self.y_, zeta=?, zeta_labels=?)
             #     #raise NotImplementedError("Result for pre_sample not available")
